# openapi.yaml
openapi: "3.1.0"
info:
  title: Stellar Wrap API
  version: "1.0.0"
  description: |
    API routes exposed by the Stellar Wrap Next.js application.

    Public endpoints:
    - **GET /api/wrapped** — returns aggregated on-chain statistics for a Stellar address.
    - **GET /api/og** — returns a 1200 × 1200 PNG share card (served from Vercel Edge Runtime).

    Notification endpoints (push + email subscriptions, GDPR data deletion):
    - **POST /api/notifications/subscribe** — store a browser push subscription.
    - **POST /api/notifications/subscribe-email** — start an email subscription (pending confirmation).
    - **GET /api/notifications/confirm-email** — confirm a pending email subscription.
    - **POST /api/notifications/unsubscribe** — unsubscribe by token or by wallet + channel.
    - **POST /api/notifications/dispatch** — triggered by a Vercel cron; fans out due notifications.
    - **GET /api/notifications/preferences/{wallet}** — read a subscription record.
    - **PUT /api/notifications/preferences/{wallet}** — update a subscription record.
    - **DELETE /api/notifications/data/{wallet}** — GDPR data deletion request.
servers:
  - url: https://stellar-wrap.vercel.app
    description: Production
  - url: http://localhost:3000
    description: Local development
security: []
paths:
...
  /api/notifications/subscribe-email:
    post:
      operationId: subscribeEmail
      summary: Start an email subscription (pending confirmation)
      description: |
        Validates an email address and wallet, writes a pending email
        subscription to KV, and sends a confirmation email with a
        confirmation link. The email is only active after the link is visited
        (`GET /api/notifications/confirm-email`).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailSubscribeRequest"
            example:
              walletAddress: "GABQUEIYD4TC2NB3IJEVAV26MVWHG6UBRCHZNHNEVOZLTQGHZ3K5ZEJS"
              email: "alice@example.com"
              periods:
                weekly: true
                monthly: true
                yearly: false
      responses:
        "200":
          description: Confirmation email sent; subscription is pending.
          content:
            application/json:
              schema:
                type: object
                required: [ok, status]
                properties:
                  ok:
                    type: boolean
                    example: true
                  status:
                    type: string
                    enum: [pending]
                    example: "pending"
              example:
                ok: true
                status: "pending"
        "400":
          description: Invalid wallet address or invalid email address.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                invalidWallet:
                  value:
                    error: "Invalid wallet address"
                invalidEmail:
                  value:
                    error: "Invalid email address"
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "Internal server error"
...
  /api/notifications/unsubscribe:
    post:
      operationId: unsubscribe
      summary: Unsubscribe from notifications
      description: |
        Two modes, mutually exclusive:
        - `{ "token": "..." }` — one-click unsubscribe using the token from the
          footer of a notification email.
        - `{ "walletAddress": "...", "channel": "push" | "email" }` — disable a
          single channel from the preference page.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UnsubscribeRequest"
            examples:
              byToken:
                value:
                  token: "f3a5c1d9e2b8476fa0b1c2d3e4f5a6b7"
              byChannel:
                value:
                  walletAddress: "GABQUEIYD4TC2NB3IJEVAV26MVWHG6UBRCHZNHNEVOZLTQGHZ3K5ZEJS"
                  channel: "email"
      responses:
        "200":
          description: Successfully unsubscribed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
              example:
                ok: true
        "400":
          description: Missing token or walletAddress+channel.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "Provide either token or walletAddress+channel"
        "404":
          description: Token not found or no subscription for the wallet.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                tokenNotFound:
                  value:
                    error: "Token not found"
                noSubscription:
                  value:
                    error: "No subscription found"
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "Internal server error"
  /api/notifications/dispatch:
    post:
      operationId: dispatchNotifications
      summary: Fan out due notifications to all subscribers
      description: |
        Evaluates which wrap periods have just started and fans out push/email
        notifications to every matching subscriber. Idempotent per
        wallet/channel/period — a dispatch log entry prevents duplicate sends.

        **Cron**: Driven hourly by a Vercel cron job
        (`vercel.json` → `0 * * * *`).

        **Authentication**: Requires `Authorization: Bearer $CRON_SECRET`.
        If `CRON_SECRET` is not configured on the server, the check is skipped.
      security:
        - cronSecretAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                periods:
                  type: array
                  description: Override which periods to dispatch instead of deriving from the current time.
                  items:
                    $ref: "#/components/schemas/WrapPeriod"
            example:
              periods:
                - monthly
      responses:
        "200":
          description: Dispatch completed.
          content:
            application/json:
              schema:
                type: object
                required: [ok, dispatched, periods]
                properties:
                  ok:
                    type: boolean
                    example: true
                  dispatched:
                    type: integer
                    minimum: 0
                    description: Number of notifications actually sent.
                    example: 42
                  periods:
                    type: array
                    description: Periods considered by this dispatch.
                    items:
                      $ref: "#/components/schemas/WrapPeriod"
                  message:
                    type: string
                    description: Present when no periods were active.
                    example: "No active periods"
              examples:
                withDispatch:
                  value:
                    ok: true
                    dispatched: 42
                    periods:
                      - monthly
                noActivePeriods:
                  value:
                    ok: true
                    dispatched: 0
                    periods: []
                    message: "No active periods"
        "401":
          description: Missing or invalid cron secret.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "Unauthorized"
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "Internal server error"
...
components:
  securitySchemes:
    cronSecretAuth:
      type: http
      scheme: bearer
      description: |
        Bearer token matching the `CRON_SECRET` environment variable. Set by the
        Vercel cron job that calls `POST /api/notifications/dispatch`.
  schemas:
    WrapPeriod:
      type: string
      enum: [weekly, monthly, yearly]
      example: monthly
    PeriodPrefs:
      type: object
      required: [weekly, monthly, yearly]
      properties:
        weekly:
          type: boolean
          example: true
        monthly:
          type: boolean
          example: true
        yearly:
          type: boolean
          example: false
...
    EmailSubscribeRequest:
      type: object
      required: [walletAddress, email, periods]
      properties:
        walletAddress:
          type: string
          pattern: "^G[A-Z2-7]{55}$"
          description: Stellar public key (56 characters, starting with "G").
          example: "GABQUEIYD4TC2NB3IJEVAV26MVWHG6UBRCHZNHNEVOZLTQGHZ3K5ZEJS"
        email:
          type: string
          format: email
          example: "alice@example.com"
        periods:
          $ref: "#/components/schemas/PeriodPrefs"
    UnsubscribeRequest:
      type: object
      description: One of `token` or `walletAddress` + `channel` (mutually exclusive).
      properties:
        token:
          type: string
          description: Unsubscribe token from a notification email.
          example: "f3a5c1d9e2b8476fa0b1c2d3e4f5a6b7"
        walletAddress:
          type: string
          pattern: "^G[A-Z2-7]{55}$"
          description: Stellar public key.
          example: "GABQUEIYD4TC2NB3IJEVAV26MVWHG6UBRCHZNHNEVOZLTQGHZ3K5ZEJS"
        channel:
          type: string
          enum: [push, email]
          description: Channel to disable on the preference page.
          example: "email"
    OkResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
          example: true
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message safe for clients.
        code:
          type: string
          description: Stable machine-readable error code.
          enum:
            - INVALID_WALLET
            - INVALID_PUSH_SUBSCRIPTION
            - INVALID_EMAIL
            - INVALID_REQUEST
            - NOT_FOUND
            - INVALID_CONFIRMATION_TOKEN
            - INVALID_UNSUBSCRIBE_TOKEN
            - UNAUTHORIZED
            - INTERNAL_ERROR
        requestId:
          type: string
          format: uuid
          description: Correlation id included for internal failures.